Add a discover route to reorganise some of the logic into more predictable components
